refactor: change structure of folders according to task
