+++ src/tracks/track.service.ts
@@ -3,8 +3,8 @@ import { database } from 'src/db/database';
 import { NotValidIdException } from 'src/errors/notValidId';
 import { Album } from 'src/types/album';
 import { Artist } from 'src/types/artist';
-import { Track } from 'src/types/track';
-import { CreateTrackDto, UpdateTrackDto } from 'src/validators/trackValidators';
+import { Track } from 'src/tracks/trackInterface';
+import { CreateTrackDto, UpdateTrackDto } from 'src/tracks/trackDtos';
 import { validate, v4 as uuidv4 } from 'uuid';
 
 @Injectable()
