+++ src/types/database.ts
@@ -1,8 +1,8 @@
 import { Album } from './album';
 import { Artist } from './artist';
 import { Favorites } from './favorites';
-import { Track } from './track';
-import { User } from './user';
+import { Track } from '../tracks/trackInterface';
+import { User } from '../users/userInterface';
 
 export interface Db {
   users: User[];
